requirements.txt
openai
flask
chromadb
python-dotenv
pytest
